tests/wpt/web-platform-tests/css/css-overflow/parsing/continue-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Overflow: parsing continue with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-overflow/#propdef-continue">
<meta name="assert" content="continue supports the full grammar 'auto | discard'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("continue", 'auto');
test_valid_value("continue", 'discard');
</script>
</body>
</html>
